[BACKPORT 2.14][#11030] xCluster,TabletSplitting: Handle master failovers during ProcessSplitTabletResult
Summary:
Changing where xCluster hooks into the tablet splitting code. Will now run ProcessSplitTabletResult after having registered the children, but before actually sending the tablet split rpcs. This ensures that in case of a master failover at any point, the tablet split will be in a pending/started state, and will thus retry via the TabletSplitManager. This removes the need of having loaders to handle any xCluster metadata updates (cdc_state table row entries for children + entries in the consumer-producer tablet mapping).

Additionally, fixing issues with setting the checkpoint for the child tablets. Changing the logic so that we now correctly set the children tablet checkpoints in cdc_state upon reading a SPLIT_OP during a GetChanges for the parent tablet. To avoid race conditions here, we check if the master has created the children entries in the cdc_state table yet, and if not, we only return the records before this opid.

Original commit: 288330db51749705d770dd9f4a66f1fc25c15283 / D17795

Test Plan:
```
ybd --cxx-test integration-tests_xcluster-tablet-split-itest --gtest_filter XClusterExternalTabletSplitITest.MasterFailoverDuringConsumerPostSplitOps
ybd --cxx-test integration-tests_xcluster-tablet-split-itest --gtest_filter XClusterExternalTabletSplitITest.MasterFailoverDuringProducerPostSplitOps
```

Reviewers: nicolas, rahuldesirazu, hsunder, asrivastava

Reviewed By: asrivastava

Subscribers: bogdan, ybase, lnguyen, hsunder

Differential Revision: https://phabricator.dev.yugabyte.com/D17838

+++ src/yb/integration-tests/cdc_test_util.cc
@@ -115,7 +115,7 @@ size_t NumProducerTabletsPolled(MiniCluster* cluster) {
 }
 
 Status CorrectlyPollingAllTablets(
-    MiniCluster* cluster, uint32_t num_producer_tablets, MonoDelta timeout) {
+    MiniCluster* cluster, size_t num_producer_tablets, MonoDelta timeout) {
   return LoggedWaitFor(
       [&]() -> Result<bool> {
         static int i = 0;
